Rename CommandClassVerion to Version
The specification using confusing language but the correct naming of
this command class is `VERSION`, which is used to get all data about
versioning including but not limited to command class versions.

I made the updates to reflect the specification better and to better
add support to the `VERSION` command class such as getting Z-Wave
software and serial API versions.