Generate zipgateway.cfg from options passed in
When having Grizzly start zipgateway (useful for Nerves systems) we now
generate the `zipgateway.cfg` file before running the binary. This is
useful for device specific information like product id to be passed into
the configuration. Also, this will allow for optioning out of the
default scripts provided. If running in a Nerves system you can write
your own scripts and put them in the `rootfs_overly` directory and
provide that path to the script in the configuration.

The base configuration generated when there are no options provided is
the basic configuration that works in the Nerves environment.