Fix when sending the same command to the same node
When sending the same command to one node from two different processes
one command would be completed and the other would be marked as
completed as well, but that was wrong. Since both commands wait for the
same the report we one only complete one and allow the other to wait for
the next report that it is waiting for to allow it to be marked as
complete.