Update DoorLockOperationReport timeout for new spec
The newly released Z-Wave specification updated the door lock operation
report timeout in minutes max minutes from `0xFC` to `0xFD`. This is a
fix in their documentation has they had stated that `253` (`0xFD`) is
valid. This should be noop for most Grizzly users and will mostly add a
greater range of support for locks that contain `0xFD` in the report.