Fix ack request handling
There are two different times we can get an `:ack_request`

1. When the `zigpateway` server sends one to ping the client
2. When the `zipgateway` server wants to ensure packet delivery to the
client.

Grizzly only handled the first one. With `zigpateway` 7.14 it appears it
sends `:ack_request` on all packets so this fix should account for the
second reason `Grizzly` might receive an `:ack_request` as part of the
Z/IP packet with a command report.