Fix connection supervisor not being ready for connection check
When the argument `:on_ready` is passed to the Grizzly supervisor it
will start a process that polls the connection to the Z-Wave controller
node until it gets a connection to let the consuming application know
that it can now safety control devices.

This moved the checker down to the last process to start to ensure that
the connection supervisor was up and running able to start new
connections to devices.