Refactor Associations Persistence
There are some logical requirements that needed to be fulfilled around
associations and the existing code did not seem handle the cases very
well. Moreover, there was a strong coupling of the file writing and
reading code with the command processing code. I pulled the coupling
apart a bit more to allow for better testing and separation of logical
units of work.