Remove sending ack response to unsolicited notifications
We generalized sending and `:ack_response` to an unsolicited
notification when though the unsolicited server is only suppose to
receive commands and not send as per the `zipgateway` documentation.
This removes doing that and tries to handle a supported extra command if
we cannot handle the extra command we assume that it is a notification
and we broadcast that to who ever is registered to list for that
notification.