Support SupervisionGet encapsulated notifications
Support when a notification is encapsulated in a `SupervisionGet`
command.

We have to inspect the inner encapsulated command and decode it. Then we
can pass back the command for the `handle_response/3` function can use
the decoded inner command for broadcasting to all the registered
listeners.

If we are unable to decode it, we will just log a warning for now as the
main mechanism for subscribing to unsolicited reports does not support
listening from binary representations yet.