Refactor extra command handling
This refactor allows for the testability of how Grizzly responses to
extra commands. Responding to extra commands correctly is important and
as such should be tested. Plus, some of the code became a bit hard to
reason about thus causing more risk to maintain.

This transforms the unsolicited server response handler to return data
structures rather than handling the broadcasting of notifications or
sending binary data to the Z-Wave network.

Also, makes an option for providing the associations server as an option
and allows for us to make testing more reproducible and limits the
amount of runtime dependencies.