Support AssociationGroupNameGet in unsolicited server
When a node requests an association group name we respond with the name
report for the lifeline association.

If the `group_id` name requested is not supported we are to respond with
the name report for the lifeline association. As of right now we are
going to support only the lifeline association so this codes will always
return with the lifeline association name.

Also, so nitpick clean up that has no effect on code functionality.