Support supervision get command fully for unsolicited server
When the unsolicited command for supervision get, we use to just notify
listeners of the internal command that was sent, now we reply back with
the `SupervisionReport` command as outlined in the Z-Wave spec.

Also, did some clean up and switched the private functions in the
`ResponseHandler` module to return the `actions` list as outlined in the
type spec.