Support multi command encapsulated in unsolicited server
This supports when the unsolicited server receives the multi command
encapsulated command.

It will inspect each command that in encapsulated.

If the command is part of the extra command classes we will pass that
long to the handler_command function for that command.

If the command is the `:alarm_report` we will still notify. We have to
provide this explicitly because we explicitly check the command against
the extra command classes so we are not blindly calling
`handler_command/2` function.

If the command does not meet either of those requirements we will tell
whoever is calling this function that they should pass it to the Z-Wave
controller (Z/IP Gateway) and send the report back that requesting node.