Support the multi channel association get, set, remove handling
This supports the multi channel association commands get, set, and
remove handling in the unsolicited server.

This assumes right now that the device using Grizzly does not support
multi channels and if I understand the specification correctly that
means we are to mirror how we support this in the non-multi channel
association command class.

There is some duplicate code here however, after more testing I will
imagine that code will be cleaned up as we get a better understanding of
the specification.