Update xref application boundaries for EEx
When updating a consuming project to Elixir 1.11.2 there were warnings
about `EEx` not being a dependency, however we make calls to `EEx` for
generating commands and command class modules.

This lets the compiler know that we actually don't depend on `EEx`
when using Grizzly, as it is only leveraged during compile time/code
generation.