Respond to version queries for extra commands
There are two situations that we need to respond to extra commands:

1. From the Z-Wave PAN
2. From a user of Grizzly

For the first case the Z-Wave Z/IP packet will be forward to the
unsolicited destination server which will have to handle send back the
version command class report.

For the second case a user might call `Grizzly.send_command/4` to query
for a command class version.

Packets are only forward to the unsolicited destination server when the
come from the Z-Wave PAN network. For a user that calls `send_command`
the Z-Wave packet will never be forwarded to due this limitation (which
appears to be in place for security reasons). Since, the information for
the version reports are hardcoded information only we just respond to
callers of `send_command` directly without any Z-Wave commands being
sent. The abstraction still behaves like the normal `send_command` to
not break API expected functionality.