Support more that just timeout errors on DTLS failures
These are due to zipgateway not being ready and they're ok. The strategy
is to retry later. This fixes crashes coming from this code that look
like this:

```
18:22:03.081 [error] GenServer Grizzly.ZIPGateway.ReadyChecker terminating
** (CaseClauseError) no case clause matching: {:error, {{:case_clause, {:error, :eaddrnotavail}}, [{Grizzly.Connections.SyncConnection, :init, 1, [file: 'lib/grizzly/connections/sync_connection.ex', line: 69]}, {:gen_server, :init_it, 2, [file: 'gen_server.erl', line: 417]}, {:gen_server, :init_it, 6, [file: 'gen_server.erl', line: 385]}, {:proc_lib, :init_p_do_apply, 3, [file: 'proc_lib.erl', line: 226]}]}}
    (grizzly 0.16.1) lib/grizzly/connections/supervisor.ex:52: Grizzly.Connections.Supervisor.do_start_connection/3
    (grizzly 0.16.1) lib/grizzly/zipgateway/ready_checker.ex:25: Grizzly.ZIPGateway.ReadyChecker.handle_continue/2
    (stdlib 3.13.2) gen_server.erl:680: :gen_server.try_dispatch/4
    (stdlib 3.13.2) gen_server.erl:431: :gen_server.loop/7
    (stdlib 3.13.2) proc_lib.erl:226: :proc_lib.init_p_do_apply/3
```

This additional error came up with the change to bind to the unsolicited
IP address. This IP address doesn't exist until zipgateway sets up the
tunnel. Previously, Erlang would use whatever IP address was available
at the time. This wasn't right, but it didn't crash with
`eaddrnotavail`. Presumably it timed out and the retry "just worked".