Process supervision encapsulated command
When a command is sent via supervision get we were not processing that
command internally. With supervision get commands we have to send the
supervision report last after doing any actions that are needed based
on the encapsulated command.

Supervision get commands only encapsulate set type of commands that do
have a report to follow. Due to this Grizzly does not need to worry
about encapsulating commands in supervision at this level as of right
now.