Add support for a database file
When using `zipgateway` version >= 7.14.2 it will rely on a database
file to save network and Z-Wave information.

You can pass the file path to the `Grizzly.Supervisor` via the
`:database_file` option but this does default to `/data/zipgateway.db`.

When migrating to a version that is >= 7.14.2 `Grizzly` will try to run
the `zgw_eeprom_to_sqlite` utility that is packaged with `zipgateway`.
The expected location for this utility is
`/usr/bin/zgw_eeprom_to_sqlite`.