Ensure that the DSK binary is 128 bits; add test
DSK.new/1 was untested. This fixes that.

The type spec for DSK.t() also requires that the binary be 128 bits
internally. This is nice since it means that DSK functions don't have to
worry about empty or short binaries. It, however, was not implemented
that way and binaries could be any even size between 0 and 16 bytes.