Add PIN handling functions to DSK module
This adds `DSK.parse_pin/1` which is a friendlier way of handling PINs
from users and APIs. This pulls in the validity checking and removes the
need to pad zeros on strings.

The other function, `DSK.to_pin_string/1` reverses the process. Most
importantly it pads PINs with leading zeros which is a super easy
mistake to make.