Refactor Meta Extensions
As SmartStart and the related meta extensions seem to be the direction
Z-Wave is going in regards to long range it seemed like a fitting time
to update the code from some of the complexities that I originally
introduced on the first implementation.

This removes all the structs for each meta extension and just allows
each extension to work like an keyword option list.

Also, this starts moving the code towards using iodata, which is more of
a long term refactor but it is nice to start introducing to the code
base when possible.