Support extended node ids in node list report
This refactors our parsing of the node list binary. This type of code is
used in various places of Grizzly so providing an internal module to
reuse will provide useful. A follow up PR will made to update the parts
of code that would use this module.

This also handles parsing the node list binary bit mask for extended
node ids in support of Z-Wave long range.