Support extended node ids in failed node list report
This change proves support for parsing and encoding extended node ids
for failed node list report. This also moves the encoding logic to the
NodeIdList module to better provide code reuse for encoding node id
lists correctly. This logic is located in other parts of the code, so
follow up changes will update the other places in the code that can use
this encoding logic.