Fix when invalid packets from from Z-Wave
Sometimes Z-Wave will send out of spec information to us. Either by our
request or by some action in the Z-Wave network that we have setup to be
notified about. Originally this was designed to crash hard but as the
use case has grown it seems better to allow the sent command to time out
while logging that Grizzly wasn't able to parse the packet coming from
the Z-Wave network.

This prevents automations getting stuck hitting the same crash for a
device that does not follow spec, thus causing Grizzly to infinitely
crash.