Fix crash when zipgateway responds with queue full
When sending messages to a wake up device, while the device is sleeping
the command gets queued in zipgateway's mailbox. The mailbox can only
contain 2000 entries by all devices on a network. If a command is sent
and is queued into the mailbox but the mailbox has hit capacity the
command is dropped and the device will never receive it.

This fixes an issue were Grizzly was not matching on this error when
receiving a Z/IP Packet from zipgateway that contains the
`:nack_queue_full` flag.