Stop and start zipgateway
With some changes to the firmware update, and to support zipgateway
backup and restore functionality later, Grizzly needs to be able to
stop and start zipgateway before some of those process are started. The
reason for this is that Grizzly will call out to tooling that will
will need to attach to the same serial port as zipgateway does to run,
so these tools and zipgateway cannot run at the same time.

I have kept these functions internal to Grizzly, via moduledoc false,
since higher level functions will be provided to interact with these
processes in a more user friend way and abstracts the idea of stopping
and starting zipgateway.