Fix get_info/2 to allow forcing a cache update
This provides better support for force a cache update when getting node
information and provides a shorter default time to refresh the cache.

Sometimes, refreshing a device's information in the cache takes more
time to complete, so this increases the timeout value to 10 seconds
rather than the default 5 seconds. If the information is cached and
there is no force update or the max age of the record is still okay the
information returns immediately and the increased timeout has no effect.