Isolate inclusion status from inclusion logic
We isolate the inclusion status from the inclusion logic process in order
to restore the inclusion process back to a working state in case of
runtime errors in the main inclusion process.

This fixes a bug where the inclusion process would be restored back to
an idle state, but the Z-Wave controller is still in inclusion mode.