Fix RSSI encoding
In #705, we fixed the decoding of the full range of expected RSSI
values, but we missed re-encoding. This is normally not a problem
because Grizzly typically only decodes RSSI_REPORTs, but this will cause
calls to `Grizzly.Trace.dump/1` raise an error if an RSSI_REPORT with an
out-of-range value exists in the trace buffer.