Handle errors raised when writing trace dumps
There is a bug in writing trace dumps when a command in the trace buffer
is not a Z/IP Packet encapsulated command. It's unclear what command(s)
are involved as the issue seems somewhat rare.

Until we can pin down the issue, this should prevent trace dumps from
crashing under the described circumstances and provide debugging
information for a future fix.