Validate command names in `Grizzly.Commands.Table`
Several commands had names that did not match the command name in the
command returned from the implementation's `new/1` function. Several
other commands were pointing to the wrong implementation module and
never would have worked properly.

This adds simple validation to `Grizzly.Commands.Table` to ensure that
(outside of a few exceptions) the command name in the table matches the
command name in the implementation module.