Fix some more RSSI_REPORT quirks
As it turns out, Z/IP Gateway will send whatever RSSI value it actually
measures rather than just what's in the spec. This means that -10dBm is
valid even though the spec says only -32..-128. When Grizzly received an
out-of-spec value for LR channels, it would return a DecodeError.
Out-of-spec values for non-LR channels would be silently dropped, which
is not desirable behavior (since you have no way to know what channel(s)
any remaining values belong to).

Instead of continuing to play whack-a-mole with parsing these values, we
are going to stop validating them entirely.

The special values 0x7D, 0x7E, and 0x7F (below sensitivity, saturated,
and not available) will be converted to atoms as before. All other
values will be treated as signed 8-bit integers. Values greater than
about -15dBm are still probably garbage, but it's up to end applications
to determine how to handle those cases.