Reset tun/tap interface by marking it down
Z/IP Gateway was failing to start up due to an issue with setting routes
on the tun/tap network interface. Setting the interface to down and
restarting the `zipgateway` process fixed this. In order to prevent this
from happening in the future, this updates the tunnel setup script to
bring the interface down first and then bring it up. Doing so should
clean up state in the Linux kernel and hopefully make setting up the
tun network more reliable. In theory, it also brings the network down if
requested to also remove routing table state.