Update RSSI to signal bar calculation
Typical "good" Z-Wave RSSI values are in the -60..-80dBm range. It's
rare to see anything below -50dBm unless the sending node's radio is at
max transmit power and the receiving node is in physical contact with
it.

On the other end of the range, most Z-Wave devices stop responding
around -100dBm (the exact value will depend on many factors, but in a
relatively quiet RF environment, this is the expected norm). Most
devices remain responsive even with an RSSI as low as -95dBm, so
dropping to 0 bars at -80dBm made very little sense.

Additionally, dBm is a logarithmic scale -- each 10dB interval
represents a 10-fold change in power level, so a 3dB decrease is
roughly equal to a halving of power.

| Bars | Previous Range | New Range     |
| ---- | -------------- | ------------- |
| 4    | >= -50dBm      | >= -78 dBm    |
| 3    | -51 – -60dBm   | -77 – -84 dBm |
| 2    | -61 – -70dBm   | -85 – -90 dBm |
| 1    | -71 – -80dBm   | -90 – -97 dBm |
| 0    | <= -81dBm      | <= -98 dBm    |