Support signed values in Thermostat Setpoint Set
The Thermostat Setpoint Set CC uses signed values in its binary
encoding, but we were encoding and decoding values as unsigned.

On top of being incorrect, the calculation for the number of bytes
needed to encode the value would raise an error for any negative number
(because log-n is undefined for values <= 0).

Co-Authored-By: Mike McCall <mike.mccall@smartrent.com>