Z/IP Gateway ready checker waits for initial node list report
Z/IP Gateway sends an unsolicited Node List Report when it is far enough
along in its initialization process to receive commands. Instead of
testing that we can merely open a connection, we'll wait for that
signal.

If we don't receive it after 60 seconds (and every 60 seconds thereafter
until success), we'll try to get the node list ourselves.