Ignore illegal values for door lock mode
Long-term, I want to remove all DecodeErrors from unexpected/illegal
enum values and instead return `:invalid` (or `:unknown` or maybe just
the original value) and then log a warning.

Ideally, we would also want a way to signal (perhaps in the
`Grizzly.ZWave.Command` struct) that we failed to parse one or more
parameters so that API consumers have a programmatic way to handle those
cases. It just seems unreasonable to throw away a frame because one byte
(or in some cases a single bit) wasn't what we expected.