Reports indicate command acknowledgement
When dealing with timeouts, it was impossible to tell whether the
outgoing command had been acknowledged by the destination node or not.
This adds an `acknowledged` flag to `Grizzly.Report` to indicate such.
This is important in determining the reason for a timeout and whether
failed command should be retried.