Z-Wave module firmware upgrade improvements (#840)
This is a major rewrite of the Z-Wave module firmware upgrade
functionality to improve the safety and reliability of the upgrade
mechanism.

New functionality:

* Upgrades can place requirements on the current firmware version
* Option to hard reset the Z-Wave module before applying an upgrade or
  after a failed upgrade (helps return the module to a clean state)
* Ability to detect when the Z-Wave module is stuck at the bootloader
  (usually due to a failed upgrade) and automatically try to use
  `zw_programmer`'s `-a` flag[^1].

The previous version required upgrade specs to specify a chip type,
which we would match against what was returned from the serial API.
In the case of a failed firmware upgrade or corrupted application on the
Z-Wave module, we will not be able to determine this. Because we can't
run this check in all cases, I decided it would be better to remove it
and require that consumers only supply firmware images that are
compatibile with the running hardware.

[^1]: At the time of writing, `zw_programmer`'s `-a` flag does not work
with 700- or 800-series modules. You will need to apply a patch that
allows use of the `-a` flag with 700/800-series modules when the `-7`
flag is supplied. Because Z/IP Gateway is not licensed for
redistribution, we cannot provide a patch file.