Close all DTLS connections on Z/IP Gateway exit
Since these are UDP connections, we have no way of knowing if the other
side has been closed until a keepalive fails.

This doesn't account for the UnsolicitedServer, but it's less of an
issue since it only deals with incoming datagrams. It's still leaking
resources and is something that sould be fixed (but separately).