Virtual devices have statically assigned IDs
Virtual device IDs are currently unstable, as they are assigned as the
count of existing virtual devices + 1. This will also cause major issues
if a virtual device process ever crashes, as it will not be able to be
reregistered without restarting all virtual devices in the process tree.

Instead, IDs should be assigned by the API consumer at the time of
registration. This will assure that `{:virtual, 1}` doesn't switch
implementations with `{:virtual, 2}` because of a change in their
startup order.

**Breaking change**