Allow creation of unnamed `AsyncConnection`s
`AsyncConnection` is a bit of a misnomer. While it does only support
sending commands asynchronously, the defining feature (and a critical
point that was never documented) is that an `AsyncConnection` is meant
to be owned and used by a single process (thus its use in inclusion and
firmware upgrades). A clearer name would be `OwnedConnection`.

The upshot of this behavior is that if two different processes were to
send a command with `mode: :async`, the first process to do so would
receive all of the status messages and the second process would get
nothing.

This adds an option to `AsyncConnection.start_link/3` to start an
unnamed connection rather than potentially reusing an existing named
connection.