Firmware update dynamic transmission delays
Firmware updates now dynamically adjust the delay between transmitting
fragments based on the active transmission speed (10ms for 100kbps and
35ms otherwise).

It is also possible to override the transmission delay in case rapid
transmissions cause issues for a particular device.

Additionally, any commands that fail during a firmware update will be
retried up to 2 times in case of nack responses due to transmission
bursts.