Revert removing the `listening?` node info property
It was removed on the basis of the Z-Wave documentation at "5.2.5.4.4. Node info cached report" describing the `List.` bit as

```
List (1 bit)The Optional Functionality bit indicates if true ( == '1') the node supports more command classes in addition to
the ones covered by the device classes listed in this message. The additional command classes follow the device class fields.
```

The official documentation is wrong here. The `List` bit does indicate whether or not the device is "mains powered" and thus constantly
listening. See "2.1.4.2. Application Specific Part" in the Z-Wave documentation for a description of the `listening` flag which value is
the same as the `List.` bit in the Node Info Cached Report command.