Fix flaky WaitReport handler test
The test tries to send two commands out of order, but sometimes due to
the two async processes not being coordinated, the commands sometimes
end up going out in order. However, the mock Z/IP Gateway server always
sends the responses out of order, so we just need to make sure that the
data is matched up correctly.