Check Z/IP Gateway database integrity at startup (#1117)
A rare bug in ZGW's data store can cause ZGW to enter an infinite loop
from which it will never recover. This is due to a missing entry in the
`endpoints` table in its database, which can be corrected by deleting
the record from the `nodes` table and allowing ZGW to re-interview the
node.