Rename `Grizzly.Commands` to `Grizzly.Requests` (#1145)
The naming of `Grizzly.Commands` makes it very easy to conflate the
modules and structures used for tracking requests to Z/IP Gateway with
the actual Z-Wave commands being sent.

To that end, "Grizzly commands" are now requests, and Z-Wave commands
remain the same.

While this does contain breaking changes, they should be fairly minor
as most of the affected code is internal to Grizzly.